requirements.txt
Flask
azure-functions
mysql-connector-python
requests